requirements.txt
asgiref==3.5.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-anymail==8.6
django-browser-reload==1.3.0
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sib-api-v3-sdk==7.4.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
tornado==6.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.1
